requirements.txt
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
psycopg2==2.9.9
PyJWT==2.8.0
pymongo==4.6.2
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.10.0
